chore: componenting header
